update CustomerListPage
- separate search in hooks
